+++ packages/react/src/index.ts
@@ -1,2 +1,2 @@
-// @paul/react - React component wrappers for the Paul Design System
-// Components will be exported from here as they are built.
+export { Button } from './Button';
+export { cx } from './cx';